renamed argument in formatMoney()

+++ src/currency/exchanger/MoneyFormatter.java
@@ -2,9 +2,9 @@ package currency.exchanger;
 
 public class MoneyFormatter
 {
-    public static String formatMoney(int integer)
+    public static String formatMoney(int money)
     {
-        String stringResult = Integer.toString(integer);
+        String stringResult = Integer.toString(money);
         StringBuilder stringBuilder = new StringBuilder(stringResult);
 
         stringBuilder.insert(stringResult.length() - 2, ".");
